requirements.txt
requests==2.27.1
pytest==7.0.1
python-dotenv==0.19.2
pytest-cov==3.0.0
pytest-xdist==2.5.0
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0
python-dateutil==2.8.2
twine
werkzeug==2.0.0
pyhackthebox==0.5.6
flask==2.0.3
# Static type checking
mypy
types-requests
types-python-dateutil


